Revert "Move geospatial headers to subfolder (#289)"
This reverts commit e0a5490f379147fb1a1d62d84143e832f61ea2a8.
